Cleaning
Moved redundant functions and fixed a bug with swap equipment swapUser and dropdown values not being aligned/cleared
